# Logging redesign (cont'd)