--- a/src/input_output/FGLog.h
+++ b/src/input_output/FGLog.h
@@ -59,7 +59,10 @@ class Element;
 // The return type of these functions is unspecified by the C++ standard so we
 // need some C++ magic to be able to overload the operator<< for these functions.
 using setprecision_t = decltype(std::setprecision(0));
+// For MSVC set_precision_t and setw_t are the same type
+#ifndef _MSC_VER
 using setw_t = decltype(std::setw(0));
+#endif
 
 enum class LogLevel {
   BULK,  // For frequent messages
@@ -121,7 +124,11 @@ public:
   FGLogging& operator<<(std::ostream& (*manipulator)(std::ostream&)) { buffer << manipulator; return *this; }
   FGLogging& operator<<(std::ios_base& (*manipulator)(std::ios_base&)) { buffer << manipulator; return *this; }
   FGLogging& operator<<(setprecision_t value) { buffer << value; return *this; }
+  // Avoid duplicate definition for MSVC for which set_precision_t and setw_t
+  // are the same type
+#ifndef _MSC_VER
   FGLogging& operator<<(setw_t value) { buffer << value; return *this; }
+#endif
   FGLogging& operator<<(const SGPath& path) { buffer << path; return *this; }
   FGLogging& operator<<(const FGColumnVector3& vec) { buffer << vec; return *this; }
   FGLogging& operator<<(LogFormat format);