--- a/src/input_output/FGfdmSocket.h
+++ b/src/input_output/FGfdmSocket.h
@@ -44,6 +44,7 @@ INCLUDES
 #if defined(_MSC_VER) || defined(__MINGW32__)
   #include <winsock.h>
   #include <io.h>
+  #undef ERROR
 #else
   #include <netdb.h>
 #endif