This PR is the continuation of the PR #1094. The replacement of `cout` and `cerr` by `FGLogging` is propagated to some additional classes.

The class `FGLogging` has also had some changes:
* It skips calling `FGLogger::Message` if its buffer is empty. This is because we cannot assume that calling `FGLogger::Message` is a cheap operation.
* `FGLogging::operator<<` is extended to all number formats, `FGColumnVector3` and some formatting from `std::ostringstream` such as `std::setw` and `std::ios_base`. This is to ensure that `FGLogging` behaves the same way as `cout` and `cerr`.
  * There is a subtlety regarding `setw_t` as Windows considers `std::setw` and `std::setprecision` to be of the same type while Linux and MacOSX don't.
* `ERROR` has to be undef'ed in `FGfdmSocket.h` because there is a clash between `ERROR` being defined in the Windows headers and `ERROR` being used in our `LogLevel` enum:
https://github.com/JSBSim-Team/jsbsim/blob/b41fcddc505baf041a3082ac3be2312e392c7792/src/input_output/FGLog.h#L62-L69

Regarding the exception management:
* When an error will throw an exception, the error level is always set to **FATAL**.
* The message that is passed to the exception `JSBSim::BaseException` is a copy of the error message so that the same message is logged and send to the exception manager. As a result the code looks like:
```c++
FGLogging log(FDMExec->GetLogger(), LogLevel::FATAL);
log << "Something terrible happened !" << endl;
throw BaseException(log.str());
```