@bcoconni 2024-08-11
> Was a bit tedious glancing through the diff given how many lines were changed, but I guess even more tedious for you to edit them all 😉

Yep ! It's either a lot of small PRs or a few big PRs. I'm not sure which option is the best. Small PRs are easier for the reviewer but less PRs is best for the author. It really is depending on which side of the fence you stand 😉