@seanmcleod 2024-08-13
Hmm, I decided to run some before versus after comparisons to double-check for any performance degradation. Basically using the same tests I did here - https://github.com/JSBSim-Team/jsbsim/pull/1094#issuecomment-2123177017 when we first discussed the logging changes.

So I built a release build before pulling these latest changes, and then another release build after pulling the latest changes. Ran each of them 3 times and averaged their runs.

```sh
$env:JSBSIM_DEBUG=0
measure-command { .\Release\JSBSim.exe --script=scripts\737_cruise.xml }
```

| Debug Level | Before (s) | After (s) |
| --- | --- | --- |
| 0 |  0.204 | 0.205 |
| 1 |  0.221 | 0.213 |
| 2 | 0.284 | 3.275 |
| 9 | 4.668 | 4.754 |

Note that for `JSBSIM_DEBUG=2` that the new build is more than 10x slower.

I did a quick comparison of the output, expecting to see maybe 10x more output, but there is very little difference.

| Redirected output file | Size | Lines |
| --- | --- | --- |
| Before | 779,302 bytes |  13,404 |
| After | 782,220 bytes | 13,400 |

While glancing through I did notice some differences in output number format.

```sh
Before:
  Mass Properties Report (English units: lbf, in, slug-ft^2)
                                      Weight    CG-X    CG-Y    CG-Z         Ixx         Iyy         Izz         Ixy         Ixz         Iyz
    Base Vehicle                       83000.0   639.0     0.0   -40.0    562000.0   1473000.0   1894000.0        -0.0      8000.0        -0.0
0   Fuel                                 10000     520     -80     -18           0           0           0
1   Fuel                                 10000     520      80     -18           0           0           0
2   Fuel                                  4000     480       0     -18           0           0           0
                                                                                                                                            
    Total:                          107000.0   610.8     0.0   -35.1    591572.3   1539552.7   1986235.4         0.0     19109.1         0.0


After:
  Mass Properties Report (English units: lbf, in, slug-ft^2)
                                      Weight    CG-X    CG-Y    CG-Z         Ixx         Iyy         Izz         Ixy         Ixz         Iyz
    Base Vehicle                         8e+04   6e+02       0  -4e+01       6e+05       1e+06       2e+06          -0       8e+03          -0
0   Fuel                                 10000     520     -80     -18           0           0           0
1   Fuel                                 10000     520      80     -18           0           0           0
2   Fuel                                  4000     480       0     -18           0           0           0
                                                                                                                                            
    Total:                             1e+05   6e+02       0  -4e+01       6e+05       2e+06       2e+06           0       2e+04           0

```