@bcoconni 2024-08-15
Thanks for the feedback :+1:

I'm actually positively surprised by the performance of the code with debug level 0 and 1 :smile: Regarding, the other levels of debug that's more what I was expecting, including for levels 0 and 1. I'd suspect that heavily calling the constructor/destructor of `FGLogging` takes a significant toll but that's only a guess.

I'll investigate and most likely submit a PR - at least to fix the numbers formats.