@seanmcleod 2024-08-15
So glancing at the constructors and destructors I noticed the call to `Flush()` was the only real work happening. So as a quick test I simply commented it out.

```c++
virtual ~FGLogging() { /*Flush();*/ }
```

And sure enough execution time dropped from 4.7s to 234ms!

Not sure we need so much flushing anyway 😉 

Anytime `std::endl` is written to the stream it generates a flush.