@bcoconni 2024-08-17
> I noticed the call to `Flush()` was the only real work happening. So as a quick test I simply commented it out.
> 
> ```c++
> virtual ~FGLogging() { /*Flush();*/ }
> ```
> 
> And sure enough execution time dropped from 4.7s to 234ms!

Given that among other things `Flush()` displays the contents of the buffer (see line 54 of the code below), this is no surprise that you see such a drop in execution time: you're basically stripping significant bits of the message :wink: 

https://github.com/JSBSim-Team/jsbsim/blob/0b99bf8b173fe21cd3a1481695ff199b5e4a4335/src/input_output/FGLog.cpp#L49-L61

> Not sure we need so much flushing anyway 😉

As I explained in the PR #1094, the `Flush()` method is meant to be an equivalent to `EOT` (End Of Transmission):
> The `Flush` method is called to indicate the "end of transmission." At this point, the class developer can format or parse the message, append a timestamp, etc., then output the finalized message via their application interface. It is in this method that the class developer may trigger the opening of a window to show an error message and request user feedback, such as "Ok" or "Cancel."

The thing is that `FGLogConsole` sends messages to `std::cout` as they are received. It does not buffer the data so, for that particular class, the `Flush()` method indeed looks a bit over engineered.

> Anytime `std::endl` is written to the stream it generates a flush.

The occurrences of `std::endl` that are passed to `FGLogging` (i.e. to its buffer) are not transfered to `FGLogConsole` (i.e. to `std::cout`) so the flush signal of `std::endl` is swallowed by `FGLogging` and never seen by `FGLogConsole`.